moved os folder to Application Support on Mac
When a Mac application is first opened, it appears to be verified, or
checked. As PiBakery contained Raspbian, it was very large, and this
check took over 8 minutes to complete. By moving the os folder out of
the .app bundle and into Application Support, PiBakery is only a few
hundred MB, so the first open only takes a few seconds.