update drivelist module
fixes bug where list of drives would not appear in Windows 10 - the
drivelist module needed to be updated.

this does not update the drivels scanner module, and therefore where
the drivels bug occurred, users will not be able to update an SD card
currently. this will be fixed in the next update